+++ Net_Core/DataAccess/Interfaces/ISecurityEntities.cs
@@ -1,6 +1,7 @@
 ﻿using GrowthWare.DataAccess.Interfaces.Base;
 using GrowthWare.Framework.Models;
 using System.Data;
+using System.Threading.Tasks;
 
 namespace GrowthWare.DataAccess.Interfaces
 {
@@ -10,13 +11,13 @@ namespace GrowthWare.DataAccess.Interfaces
         /// Deletes the registration information given a security entity sequence id.
         /// </summary>
         /// <param name="securityEntitySeqId"></param>
-        void DeleteRegistrationInformation(int securityEntitySeqId);
+        Task DeleteRegistrationInformation(int securityEntitySeqId);
 
         /// <summary>
         /// Returns all rows from [ZGWSecurity].[Registration_Information]
         /// </summary>
         /// <returns></returns>
-        DataTable GetRegistrationInformation();
+        Task<DataTable> GetRegistrationInformation();
 
         /// <summary>
         /// Retrieves all Security Entities as a data table.
@@ -49,13 +50,13 @@ namespace GrowthWare.DataAccess.Interfaces
         /// </summary>
         /// <param name="profile">MSecurityEntity</param>
         /// <remarks></remarks>
-        int Save(MSecurityEntity profile);
+        Task<int> Save(MSecurityEntity profile);
 
         /// <summary>
         /// Saves the registration information to [ZGWSecurity].[Registration_Information].
         /// </summary>
         /// <param name="profile"></param>
         /// <returns>DataRow</returns>
-        DataRow SaveRegistrationInformation(MRegistrationInformation profile);
+        Task<DataRow> SaveRegistrationInformation(MRegistrationInformation profile);
     }
 }